fix(deleteSS): add byPass command

+++ commands/deleteSmartStructure.js
@@ -1,5 +1,6 @@
 const gulp = require("gulp");
 const signale = require("signale");
+const inquirer = require("inquirer");
 
 const { deleteSmartStructure } = require("../tasks/deleteSmartStructure");
 
@@ -31,21 +32,48 @@ exports.builder = {
     alias: "c",
     type: "string",
     required: true
+  },
+  byPass: {
+    description: "bypass confirmation message",
+    type: "boolean",
+    default: false
   }
 };
 
-exports.handler = function(argv) {
+exports.handler = async function(argv) {
   try {
-    if (argv.name) {
-      signale.info(`Deleting ${argv.name} smart structure...`);
+    if (argv.byPass) {
+      deleteSmartStructure(argv);
+      const task = gulp.task("deleteSmartStructure");
+      return task().then(response => {
+        signale.timeEnd("deleteSmartStructure");
+        signale.info("\n" + response.messages.messages.join("\n"));
+        signale.success("Smart structure has been deleted");
+      });
+    } else {
+      inquirer
+        .prompt({
+          type: "confirm",
+          name: "confirmDelete",
+          message: `You are about to delete ${
+            argv.name
+          } smart structure. Are you sure you want to delete this?`
+        })
+        .then(confirmation => {
+          if (confirmation.confirmDelete && !argv.byPass) {
+            if (argv.name) {
+              signale.info(`Deleting ${argv.name} smart structure...`);
+            }
+            deleteSmartStructure(argv);
+            const task = gulp.task("deleteSmartStructure");
+            return task().then(response => {
+              signale.timeEnd("deleteSmartStructure");
+              signale.info("\n" + response.messages.messages.join("\n"));
+              signale.success("Smart structure has been deleted");
+            });
+          }
+        });
     }
-    deleteSmartStructure(argv);
-    const task = gulp.task("deleteSmartStructure");
-    return task().then(response => {
-      signale.timeEnd("deleteSmartStructure");
-      signale.info("\n" + response.messages.messages.join("\n"));
-      signale.success("Smart structure has been deleted");
-    });
   } catch (e) {
     signale.timeEnd("deleteSmartStructure");
     signale.error(e);
